requirements.txt
flask
gunicorn
flask_cors
PILLOW
torch
torchvision